requirements.txt
uvicorn
pydantic
python-dotenv
passlib[bcrypt]
PyJWT
mysql-connector-python
requests
pandas
pymysql
gunicorn==22.0.0
uvicorn==0.23.2
fastapi==0.111.1
psycopg2-binary
SQLAlchemy==2.0.31
sqlmodel==0.0.20
celery
azure-storage-blob
redis
bcrypt
librosa
numpy
soundfile
noisereduce
webrtcvad-wheels
torch
whisperx
pandas
docx
scipy
pyannote.audio